Initial Commit
Currently contains basic html and style sheet, containing the Simon game
and buttons for start and restart. config.js is a module containing with
some of the game variables. More will need to be added. simonGame.js
contains some function to generate string of colors but that's about it
so far